Move DatePicker to FilterDateRange and comments
