+++ services/app-web/src/components/FilterAccordion/FilterDateRange/_DatePicker/DatePicker.test.tsx
@@ -17,7 +17,7 @@ import {
   VALIDATION_MESSAGE,
 } from './constants'
 
-describe('DatePicker component', () => {
+describe('_DatePicker component', () => {
   beforeEach(() => {
     jest.clearAllMocks()
   })
